Refactored the Boat system.
Before, the Boat handled both Decks and Rooms. Now, Boat only handles Decks and Decks handles Rooms instead.

Boat now contains an EventLog, and handles it for the user.

ActiveEvent is not used anymore. Rooms now have a RoomLog that keeps track of which events that is active in the room.

Sensors doesn't handle the link between the Room and the EventLog anymore, this job is done by the RoomLog instead.

LogEvent now keeps track itself it it is active or not.